Use boto3 to get full file list instead of HTTPS API
We can get the entire recursive filesystem tree from S3 easily, so we just cache
the entire tree locally and just query the local copy. Doing this with the HTTPS
API doesn't work, however, as it only returns the first 9999 results for a
directory listing, so it doesn't include most packages. With boto3, we can get
all entries.